+++ lib/math_expressions.dart
@@ -27,15 +27,23 @@
 /// expressions.
 ///
 /// The libary supports the three basic data types [Number], [Interval] and
-/// [Vector]. It includes a [Parser] to create [Expression]s from Strings.
+/// [Vector]. It includes the [GrammarParser] to create [Expression]s from
+/// string inputs.
+///
+/// For backwards-compatibility the legacy [ShuntingYardParser] is also
+/// included. Users are encouraged to switch to [GrammarParser] as it will
+/// not receive new features and bug fixes.
 library math_expressions;
 
 import 'dart:math' as math;
 
 import 'package:vector_math/vector_math.dart' show Vector2, Vector3, Vector4;
+import 'package:petitparser/petitparser.dart' hide Parser;
+import 'package:petitparser/petitparser.dart' as pp show Parser;
 
 part 'src/algebra.dart';
 part 'src/expression.dart';
 part 'src/functions.dart';
 part 'src/parser.dart';
+part 'src/parser_petit.dart';
 part 'src/evaluator.dart';